feat(deployment): Provide a dev service that starts a splunk test container
It is disabled by default for compatibility with runtime deactivation.
It can only be activated when running in Dev mode, or in QuarkusTests, through `quarkus.log.handler.splunk.devservices.enabled=true`.
It takes roughly 30s to start, all the more reason to not activate it by default.

A new test-utils deliverable is provided for compatibility with QuarkusIntegrationTests where injection of configuration is not possible.
It implements the `DevServiceContext.ContextAware` interface.

The dev service is activated in the integration-test.
Two new tests in deployment force the activation of the dev service as well.

Named handlers are supported through build time configuration:

```properties
quarkus.log.handler.splunk.devservices.plug-named-handlers.myhandlername=true
```

This is the first step towards a DevUI insert with the splunk UI inside.

Delivers parts of #240