feat(dev): Dev UI link to splunk UI
Integrating Splunk UI in the dev UI for development launch mode.
Problem, the Splunk UI is exposed with a X-Frame-Options sameorigin, which blocks the iframe, cause management port and the Splunk UI port differ.
Fortunately there is a `doNotEmbed` option on the Page builder, which makes opening the Splunk UI link in a new tab of the user's browser.
User and password to use are appended to the lin.